+++ test/integration/clusterio.js
@@ -7,7 +7,7 @@ const parallel = require("mocha.parallel");
 
 const link = require("lib/link");
 
-const { slowTest, get, exec, sendRcon, getControl, controlConfigPath, instancesDir } = require("./index");
+const { slowTest, get, execCtl, sendRcon, getControl, instancesDir } = require("./index");
 
 
 describe("Integration of Clusterio", function() {
@@ -30,62 +30,62 @@ describe("Integration of Clusterio", function() {
 
 
 	describe("clusterctl", function() {
-		describe("list-slaves", function() {
+		describe("slave list", function() {
 			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} slave list`);
+				await execCtl("slave list");
 			});
 		});
-		describe("list-instances", function() {
+		describe("instance list", function() {
 			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} instance list`);
-			});
-		});
-
-		describe("create-instances", function() {
-			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} instance create test --id 44`);
+				await execCtl("instance list");
 			});
 		});
 
-		describe("assign-instance", function() {
+		describe("instance create", function() {
+			it("runs", async function() {
+				await execCtl("instance create test --id 44");
+			});
+		});
+
+		describe("instance assign", function() {
 			it("creates the instance files", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} instance assign test 4`);
+				await execCtl("instance assign test 4");
 				assert(await fs.exists(path.join(instancesDir, "test")), "Instance was not created");
 			});
 		});
 
-		describe("create-save", function() {
+		describe("instance create-save", function() {
 			it("creates a save", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance create-save test`);
+				await execCtl("instance create-save test");
 			});
 		});
 
-		describe("export-data", function() {
+		describe("instance export-data", function() {
 			it("exports the data", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance export-data test`);
+				await execCtl("instance export-data test");
 				assert(await fs.exists(path.join("static", "export", "locale.json")), "Export was not created");
 			});
 		});
 
-		describe("start-instance", function() {
+		describe("instance start", function() {
 			it("starts the instance", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance start test`);
+				await execCtl("instance start test");
 				// TODO check that the instance actually started
 			});
 		});
 
-		describe("send-rcon", function() {
+		describe("instance send-rcon", function() {
 			it("sends the command", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance send-rcon test test`);
+				await execCtl("instance send-rcon test test");
 				// TODO check that the command was received
 			});
 		});
 
-		describe("set-instance-config-prop", function() {
+		describe("instance config set-prop", function() {
 			it("applies factorio settings while running", async function() {
 				slowTest(this);
 
@@ -159,7 +159,7 @@ describe("Integration of Clusterio", function() {
 				for (let [prop, value, ,] of testConfigs) {
 					value = `"'${JSON.stringify(value).replace(/"/g, process.platform === "win32" ? '""' : '\\"')}'"`;
 					let args = `test factorio.settings ${prop} ${value}`;
-					await exec(`node clusterctl --config ${controlConfigPath} instance config set-prop ${args}`);
+					await execCtl(`instance config set-prop ${args}`);
 				}
 
 				// Do this afterwards to leave enough to time for the
@@ -170,90 +170,134 @@ describe("Integration of Clusterio", function() {
 			});
 		});
 
-		describe("stop-instance", function() {
+		describe("user set-admin/whitelisted/banned", function() {
+			async function getUsers() {
+				let result = await link.messages.listUsers.send(getControl());
+				return new Map(result.list.map(user => [user.name, user]));
+			}
+
+			let lists = ["admin", "whitelisted", "banned"];
+			it("should add and remove the given user to the list", async function() {
+				slowTest(this);
+				await link.messages.createUser.send(getControl(), { name: "list_test" });
+				let user = (await getUsers()).get("list_test");
+				for (let list of lists) {
+					assert.equal(user[`is_${list}`], false, `unexpected ${list} status`);
+					await execCtl(`user set-${list} list_test`);
+				}
+				user = (await getUsers()).get("list_test");
+				for (let list of lists) {
+					assert.equal(user[`is_${list}`], true, `unexpected ${list} status`);
+					let remove = { admin: "--revoke", whitelisted: "--remove", banned: "--pardon" }[list];
+					await execCtl(`user set-${list} ${remove} list_test`);
+				}
+				user = (await getUsers()).get("list_test");
+				for (let list of lists) {
+					assert.equal(user[`is_${list}`], false, `unexpected ${list} status`);
+				}
+			});
+			it("should not create the user if not instructed to", async function() {
+				for (let list of lists) {
+					try {
+						await execCtl(`user set-${list} no_create_test`);
+					} catch (err) { /* ignore */ }
+				}
+				let user = (await getUsers()).get("no_create_test");
+				assert.equal(user, undefined, "user was unexpectedly created");
+			});
+			it("should create the user if instructed to", async function() {
+				for (let list of lists) {
+					await execCtl(`user set-${list} --create test_create_${list}`);
+					let user = (await getUsers()).get(`test_create_${list}`);
+					assert.equal(user && user[`is_${list}`], true, `user not created and added to ${list}`);
+				}
+			});
+		});
+
+		describe("instance stop", function() {
 			it("stops the instance", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance stop test`);
+				await execCtl("instance stop test");
 				// TODO check that the instance actually stopped
 			});
 		});
 
-		describe("delete-instance", function() {
+		describe("instance delete", function() {
 			it("deletes the instance", async function() {
 				slowTest(this);
-				await exec(`node clusterctl --config ${controlConfigPath} instance delete test`);
+				await execCtl("instance delete test");
 				assert(!await fs.exists(path.join(instancesDir, "test")), "Instance was not deleted");
 			});
 		});
 
-		describe("list-permissions", function() {
+		describe("permission list", function() {
 			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} permission list`);
+				await execCtl("permission list");
 			});
 		});
 
-		describe("list-roles", function() {
+		describe("role list", function() {
 			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} role list`);
+				await execCtl("role list");
 			});
 		});
 
-		describe("create-role", function() {
+		describe("role create", function() {
 			it("should create the given role", async function() {
 				let args = "--description \"A temp role\" --permissions core.control.connect";
-				await exec(`node clusterctl --config ${controlConfigPath} role create temp ${args}`);
+				await execCtl(`role create temp ${args}`);
 				let result = await link.messages.listRoles.send(getControl());
 				let role = result.list.find(role => role.name === "temp");
 				assert.deepEqual(role, { id: 5, name: "temp", description: "A temp role", permissions: ["core.control.connect"] });
 			});
 		});
 
-		describe("edit-role", function() {
+		describe("role edit", function() {
 			it("should modify the given role", async function() {
 				let args = "--name new --description \"A new role\" --permissions";
-				await exec(`node clusterctl --config ${controlConfigPath} role edit temp ${args}`);
+				await execCtl(`role edit temp ${args}`);
 				let result = await link.messages.listRoles.send(getControl());
 				let role = result.list.find(role => role.name === "new");
 				assert.deepEqual(role, { id: 5, name: "new", description: "A new role", permissions: [] });
 			});
 		});
 
-		describe("delete-role", function() {
+		describe("role delete", function() {
 			it("should modify the given role", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} role delete new`);
+				await execCtl("role delete new");
 				let result = await link.messages.listRoles.send(getControl());
 				let role = result.list.find(role => role.name === "new");
 				assert(!role, "Role was not deleted");
 			});
 		});
 
-		describe("list-users", function() {
+		describe("user list", function() {
 			it("runs", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} user list`);
+				await execCtl("user list");
 			});
 		});
 
-		describe("create-user", function() {
+		describe("user create", function() {
 			it("should create the given user", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} user create temp`);
+				await execCtl("user create temp");
 				let result = await link.messages.listUsers.send(getControl());
 				let user = result.list.find(user => user.name === "temp");
 				assert(user, "user was not created");
 			});
 		});
 
-		describe("edit-user-roles", function() {
+		describe("user set-roles", function() {
 			it("should set the roles on the user", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} user set-roles temp Admin`);
+				await execCtl("user set-roles temp Admin");
 				let result = await link.messages.listUsers.send(getControl());
 				let user = result.list.find(user => user.name === "temp");
 				assert.deepEqual(user.roles, [0]);
 			});
 		});
 
-		describe("delete-user", function() {
+		describe("user delete", function() {
 			it("should delete the user", async function() {
-				await exec(`node clusterctl --config ${controlConfigPath} user delete temp`);
+				await execCtl("user delete temp");
 				let result = await link.messages.listUsers.send(getControl());
 				let user = result.list.find(user => user.name === "temp");
 				assert(!user, "user was note deleted");